backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-jose[cryptography]
passlib[argon2]
argon2-cffi
python-dotenv
